pyproject.toml
[project]
name = "TranscribeTools"
version = "0.5.5"
description = "Transcribe all soundfiles in a folder using Whisper"
readme = "README.md"
requires-python = ">3.10" # see Whisper docs needs <= 3.12 but uv ignore upper bound
dependencies = [
    "attrs>=24.2.0",
    "openai>=1.59.4",
    "openai-whisper>=20240930",
    "pymsgbox>=1.0.9",
    "result>=0.17.0",
    "rich-click>=1.8.5",
    "soundfile>=0.13.0",
    "tiutools>=0.4.1",
    "toml>=0.10.2",
    "transformers>=4.49.0",
]
[project.scripts]
# commandline tool, installable
transcribefolder = "transcribetools.transcribe_folder:cli"

# below needed to run locally
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "deepl>=1.21.1",
    "pytest>=8.3.4",
]

[tool.uv.sources]
tiutools = { path = "../tiutools", editable = true }
